FOV (wip)

+++ include/objects.h
@@ -6,7 +6,7 @@
 /*   By: rubennijhuis <rubennijhuis@student.coda      +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/04/24 19:24:23 by rubennijhui   #+#    #+#                 */
-/*   Updated: 2022/05/29 16:24:04 by rubennijhui   ########   odam.nl         */
+/*   Updated: 2022/05/30 14:14:21 by jobvan-d      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -98,7 +98,7 @@ typedef struct s_camera
 {
 	t_vec3f		position;
 	t_vec3f		orientation;
-	uint32_t	fov;
+	float		fov;
 }	t_camera;
 
 typedef struct s_ambient_light
